🎈 perf(system): Multiple updates
Refactor model buffers controling

Rename `check_avaliability()` to `check_and_preprocess(args)` for transforming buffers if needed.

Change defaults of `external_model_parmas_file` from `""` to `null`.

Change the extension name of log file. From `.html` to `.log` and remove the `_log` suffix from naming rule.

Improve some log output

Add some funciton docs.

Remove the switch of `check_convergence()`.

+++ src/client/fedfomo.py
@@ -96,4 +96,4 @@ class FedFomoClient(FedAvgClient):
                         torch.stack(clients_model_params, dim=-1).to(self.device) * W,
                         dim=-1,
                     )
-                    param.data = aggregated.to(param.dtype)
+                    param.data = aggregated
